fix marking stores correctly
The stores were not being marked correctly since a variable was not
considered live in at a block if its first use was in a phi node. To
solve this problem I rewrote the LiveIn name space to use the original
live range namespace.

Also changed the way defs are collected when marking stores. We need
to collect all the names that a def could be seen under in the live
range. This is done by following defs to phi nodes, adding the phi
node name and following those names to any other phi nodes. We need to
do this so that we only insert stores at basic blocks where the def in
the live range reaches.

darcs-hash:20060811160021-22c1e-945dd65d1f862c12c40afe255d3f81ffc5d19d97