Run all optimizations in qt and rt
Changed the test scripts to run the same optimizations that are used
when collecting stats. Also added a note about the problems running st
on my mac. It needs to be run on godzilla because of problems with
g77.

darcs-hash:20070109222032-22c1e-9a85caba42a9493fba1749df1530e50290d1759c