fixed bug with accidental deletion from regMap
I was accidentally deleting an entry in the regMap if the live range
had been allocated to a temporary register in a previous block. The
solution was to set the forLRID field of the ReservedReg* structure to
be NO_LRID so that the entry would not be deleted when that reg was
allocated to a new live range in the next block.

darcs-hash:20070123165310-22c1e-8827e2aa4f331d9285e81224a47c2278a79d084b