bugfix: 002 - not selecting split point correctly
You are not selecting the first liveunit that starts with a
definition. You will choose the last unit that starts with a
definition unless you find one that is an entry point. You should
really only take the first one that starts with a def if you can not
find one that is an entry point.

darcs-hash:20070126031822-22c1e-285f1525f1940d02dc6ba6683a90780a6219055b