bugfix: 001 - closed no issue
You do not update the forbidden list for the new live range (newlr) as
you split it off from the original live range. In the function
LiveRange_IncludeInSplit you check to see if a unit can be added by
checking to see if the union of the forbidden list for newlr and the
colors used in the unit's block make a completely full set. If the set
is not full you say it is ok to add the unit, but then you never
update the forbidden set. You should be updating the forbidden set to
include the colors used in that block when you add the unit.

note:
I think this can be solved easily by updating the forbidden set in the
TransferLiveUnit() function. It looks like it is only called from the
split function and could do the job easily.

update:
Well actually you do update the forbidden set in the
LiveRange_AddBlock() function. This is called from the AddLiveUnit()
function which is called from TransferLiveUnit which is called from
the Split() function. So no bug.

darcs-hash:20070126044750-22c1e-4b6567042b2b195dafcdd05cc00b929cf6a3e998