Implemented code that can split an edge
This change allows splitting an edge with a new basic block. This is
needed for optimized placement of loads and stores in the chow
allocator.

rt was modified to allow splitting as a prepass and clean up some of
the tests that do not work on the mac because of link errors

darcs-hash:20070130202621-22c1e-c1797708c2e15287b8fc83e866f1a78b28a71a04