added option for enhanced code motion
Fixed a bug that was causing us to generate incorrect code. We were
moving a load/store regardless of whether it had any preds/succs in
the live range. This caused us to do a load before a store when a
split live range was connected by load store (i.e. both were allocated
a register and so had store/load at end/beginning of the two split
live ranges).

Now we do not move the load/store unless there is at least one
pred/succ that is in our live range. We can do better than this by
connecting those live ranges with a register to register copy. That is
the enhanced version which is stubbed but not yet implemented.

darcs-hash:20070214175644-22c1e-02cf2d6bea424f2700352e829395a28a8cc55dbc