was not evicting registers when needed for FRAME and JSR
forgot to remove a return that was cutting out some code that was doing the eviction. instead it was just using a fresh variable when it needed to evict. changed to actually evict an allocated register.

this revealed a bug with eviction where a register may not be allocated so it was not finding one to evict. this can happen when there is no advantage to allocating the register in a block so it goes unused. fixed bug to detect this case and use an unallocated register.

darcs-hash:20070404152210-22c1e-ab1f3933852ab2f0b84049a2d71a5db988028305