fixed assignment to evict registers for FRAME/JSR
This patch is the culmination of many hours of debugging. I discovered
that part of the code was commented out so that it was never actually
evicting registers for FRAME and JSR statements. Adding eviction in
caused some problems. Evicting a register means taking a live range
that was allocated a register in this block and storing it before the
JSR and then loading it after the JSR.

The problem is that if you generate a store for a live range before
using its register you may store an incorrect value since the register
could be used to hold a definition a value for the live range in that
block and storing will put an incorrect value into memory.

So what I do is to look down the instructions in the block and check
to see if it is killed before it is used or reaches the end of block.
If it is killed then no store is inserted.

darcs-hash:20070406013148-22c1e-e67a9033964f02a97a1e737da18cb9f90d2be755