fixed problem with inserting stores for evicted registers
When finding the live range that was being evicted we were doing a
linear search of the live ranges. If the live range we found was the
original live range and not the actual live range that contained the
current block then we would not find the live unit correctly and would
die. To solve this we made a inverse color map that maps blocks X
register classs X color to live range. We use this to find which live
range is being evicted.

We also solved a problem where the regMap was not being updated
correctly due to not tracking evicted register that did not cause a
load. We now track those evictions and just don't generate a load for
those registers that did not have a live range evicted.

darcs-hash:20070406071657-22c1e-74de03a7eb4ee36d7d58548b8053b5ea03bee3b8