add initialization for cfg_tools in MoveLoadsAndStores
need to inialize cfg_tools module. normally this happens in
InitCleaver(), but if not cleaving blocks it needs to happen here.
There is still a problem with initalizing it twice, but it turns out
for cfg_tools that doesn't matter. blech anyway.

darcs-hash:20070407194501-22c1e-81f33638c478c83b95800501e125b498725bee6f