fixed rematerialization to load from correct live range
The problem with the implementation was that when we assign registers
we look up the live range based on the orig_lrid. This means that the
original live range is checked for rematerialization settings and none
of the split live ranges will ever be correctly checked. We need a way
to look at the actual live range when generating loads and stores.

The idea is to get rid of the color mapping and replace it with a map
from orig_lrid x block --> current_lr. Then the color can be grabbed
from the field in the current_lr.

what about the reverse color map? it is ok since it already uses the
actual lrid, not the orig_lrid.

darcs-hash:20070517210823-22c1e-966433dce0001d103520de823f4913efea8ce6e3