implemented live range trimming
Attempts to trim away blocks that serve no purpose in the live range.
these blocks come up after splitting a live range. there can be
dangling blocks that are not part of a path that reaches a use or a
def and thus serve no purpose.  When done splitting a live range and
it is time to trim

To Trim:
use worklist
add any node where there is a use or a def.
while worklist not empty
  remove node
  look at predecessors
    if predecessor is in LR and not yet marked
      mark predecessor
      put predecessor on worklist

go through all nodes and remove any node that is not marked

note: we may leave some unconstrained lrs on the constrained list
since we are deleting some live units from existence. this is not
incorrect, but is not optimal.

darcs-hash:20070521232905-22c1e-018049888c619fd638399ee3a7412bf6f876748c