fixed problem with not freeing tmp registers
after changing it so that the tmp register map is not reset if the
block flows only to one successor there was a problem that a live
range could stay in a temp register until it reached a block
where it was actually allocated which caused problems when using the
minimum number of registers. changed the code to free up temp
registers if their live range is allocated in the successor block.

darcs-hash:20070601043611-22c1e-b7d3381127adb3f1a52fbd893f5453d246e24c31