added new heuristic and updated previous split heuristic
the new heuristic looks at the number of colors left if a block is
added and does not add the block if the number of colors is too few.
the change to the previous heuristic is how the number of neighbors is
computed if a block is added during splitting. also fixed a
calculation error for the number of new neighbors.

darcs-hash:20070601215901-22c1e-5aa778e9d2e64d554b8c65308fd138804bf6d65d