fixed bug with computing which live ranges are constrained
was incorrectly computing IsConstrained so that a live range was
getting moved to the unconstrained list when it actually was
constrained. Ended up reworking the IsConstrained function to use a
weighted neighbor count of neighbors and also divide the total
registers by the width of the current live range.

darcs-hash:20070705182733-22c1e-100eebb2dd1233aa3822951267057c04fb551b52