add a unit map to each live range mapping blocks to units
this was done to get rid of the linear lookup that was happening in
the LiveUnitForBlock() function. This gave a nice performance
increase, with twldrv.i going from 20min to 5min in one case on block
size of 2.

darcs-hash:20070727032136-22c1e-613f92cbbb21c9a4d68ccbaef97b8f84a34469e2