fixed bug with putting multiple stores on an edge
this was happening when putting a store of a double on to an edge
because the store was placed for both tmpRegs that contained part of
the double. the fix was to increment the iterator by the regwidth so
that only the one store would be placed.

darcs-hash:20070809162327-22c1e-41b051a9c857a840b6e87ea593a2b2186f0758ac