fixed problem with not finding expected load
we had been placing a copy on the successor edge even if the live
range was not live in on the successor. because the live range was not
live in there was no load and so the assert was failing. we fixed this
by not placing a copy if the live range is not live in, which is the
correct thing to do.

darcs-hash:20070810164842-a30e5-2a154d3518c2031459b411233bfdf7f5eac5895f