made sure that enhanced promotion requires code motion
Enhanced register promotion should always require enhanced code
motion. The reason is that if we try to insert a copy on an edge of a
single successor block we need to delete the load from that edge
(because there is no store there so it could load from junk memory).
So we tie enhanced promotion to require enhanced code motion.

darcs-hash:20070813145752-22c1e-71ef56afd6e0562295193759444b2137a1b8f845